arreglando ciertos bugs

+++ hostal_nomada/database/migrations/2024_10_25_133952_create_users_table.php
@@ -15,11 +15,11 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('last_name');
-            $table->integer('identification_number');
+            $table->bigInteger('identification_number');
             $table->unsignedBigInteger('nationality_id');
             $table->date('birthday');
             $table->unsignedBigInteger('rol_id');
-            $table->integer('phone')->nullable();
+            $table->bigInteger('phone')->nullable();
             $table->string('email')->unique();
             $table->string('password');
             $table->timestamp('email_verified_at')->nullable();
